Feat #4 : sitemap 정리 및 utils 구분 (리팩토링)
