+++ index.html
@@ -41,7 +41,7 @@
         </div>
         <a href="/producer">PRODUCER</a>
         <a href="/audition">AUDITION</a>
-        <div class="material-icons menu">menu</div>
+        <div id="sitemap" class="material-icons menu">menu</div>
       </nav>
     </header>
     <div id="app"></div>
