+++ src/app.js
@@ -1,56 +1,19 @@
 import Router from '@routes/router';
-
-import { $ } from './utils/querySelector';
-
-const LOCAL_URL = 'http://localhost:5173/';
+import useDropDown from '@utils/useDropDown';
+import useSPA from '@utils/useSPA';
+import Sitemap from '@components/sitemap';
 
 class App {
   constructor($container) {
     this.$container = $container;
 
-    const closedAllDropDown = () => {
-      document.querySelectorAll('.dropDown').forEach((event) => {
-        if (!event.classList.contains('closed')) {
-          event.classList.add('closed');
-          setTimeout(() => {
-            event.classList.add('hidden');
-          }, 300);
-        } else {
-          event.classList.remove('hidden');
-        }
-      });
-    };
-
     /* SPA */
-    $('.link').addEventListener('click', (e) => {
-      const target = e.target.closest('a');
-      if (!(target instanceof HTMLAnchorElement)) return;
-
-      e.preventDefault();
-      if (String(target).replace(LOCAL_URL, '/') === location.pathname)
-        history.replaceState('', '', target);
-      else history.pushState('', '', target);
-
-      closedAllDropDown();
-      dispatchEvent(new Event('popstate'));
-    });
-
+    useSPA();
     /* DropDown */
-    document.querySelectorAll('.dropDownTitle').forEach((event) => {
-      event.addEventListener('click', (e) => {
-        const { target } = e;
-        if (!(target instanceof HTMLParagraphElement)) return;
-
-        if ($(`.dropDown${target.id}`).classList.contains('closed')) {
-          closedAllDropDown();
-          $(`.dropDown${target.id}`).classList.toggle('closed');
-        } else {
-          closedAllDropDown();
-        }
-      });
-    });
+    useDropDown();
 
     new Router($container);
+    new Sitemap(document.getElementById('sitemap'));
   }
 }
