+++ src/pages/producer/index.js
@@ -1,5 +1,5 @@
 import './style.scss';
-import { $ } from '@utils/querySelector';
+import { $ } from '@utils/useQuerySelector';
 
 class ProducerPage {
   constructor($container) {
